refactor: split scanner osv helpers
